#66 refactor : CategoryFactory 빈 등록
